Move helpers and tests for an actual contract descriptor (vs mock memory/target) into separate files and subfolder
