+++ src/native/managed/cdacreader/tests/MockMemorySpace.cs
@@ -15,18 +15,10 @@ namespace Microsoft.Diagnostics.DataContractReader.UnitTests;
 /// Helper for creating a mock memory space for testing.
 /// </summary>
 /// <remarks>
-/// Use MockMemorySpace.CreateContext to create a mostly empty context for reading from the target.
-/// Use MockMemorySpace.ContextBuilder to create a context with additional MockMemorySpace.HeapFragment data.
+/// Use MockMemorySpace.Builder to create a context with MockMemorySpace.HeapFragment data.
 /// </remarks>
 internal unsafe static partial class MockMemorySpace
 {
-    // These addresses are arbitrary and are used to store the contract descriptor components.
-    // They should not overlap with any other heap fragment addresses.
-    private const ulong ContractDescriptorAddr = 0xaaaaaaaa;
-    // TODO: remove the references to these from TargetTestHelpers.ContractDescriptorFill
-    internal const uint JsonDescriptorAddr = 0xdddddddd;
-    internal const uint ContractPointerDataAddr = 0xeeeeeeee;
-
     internal struct HeapFragment
     {
         public ulong Address;
@@ -39,14 +31,8 @@ internal unsafe static partial class MockMemorySpace
     /// </summary>
     internal class Builder
     {
-        private bool _created = false;
         private readonly List<HeapFragment> _heapFragments = new();
         private readonly List<BumpAllocator> _allocators = new();
-
-        private IReadOnlyCollection<string> _contracts;
-        private IDictionary<DataType, Target.TypeInfo> _types;
-        private IReadOnlyCollection<(string Name, ulong? Value, uint? IndirectIndex, string? TypeName)> _globals;
-        private IReadOnlyCollection<ulong> _indirectValues;
 
         private TargetTestHelpers _targetTestHelpers;
 
@@ -67,49 +53,8 @@ internal unsafe static partial class MockMemorySpace
             throw new InvalidOperationException($"No fragment includes addresses from 0x{address:x} with length {length}");
         }
 
-        // TODO: contracts with versions
-        public Builder SetContracts(IReadOnlyCollection<string> contracts)
-        {
-            if (_created)
-                throw new InvalidOperationException("Context already created");
-            _contracts = contracts;
-            return this;
-        }
-
-        public Builder SetTypes(IDictionary<DataType, Target.TypeInfo> types)
-        {
-            if (_created)
-                throw new InvalidOperationException("Context already created");
-            _types = types;
-            return this;
-        }
-
-        public Builder SetGlobals(IReadOnlyCollection<(string Name, ulong Value, string? TypeName)> globals)
-        {
-            if (_created)
-                throw new InvalidOperationException("Context already created");
-            if (_globals != null)
-                throw new InvalidOperationException("Globals already set");
-            _globals = globals.Select(g => (g.Name, (ulong?)g.Value, (uint?)null, g.TypeName)).ToArray();
-            _indirectValues = null;
-            return this;
-        }
-
-        public Builder SetGlobals(IReadOnlyCollection<(string Name, ulong? Value, uint? IndirectIndex, string? TypeName)> globals, IReadOnlyCollection<ulong> indirectValues)
-        {
-            if (_created)
-                throw new InvalidOperationException("Context already created");
-            if (_globals != null)
-                throw new InvalidOperationException("Globals already set");
-            _globals = globals;
-            _indirectValues = indirectValues;
-            return this;
-        }
-
         public Builder AddHeapFragment(HeapFragment fragment)
         {
-            if (_created)
-                throw new InvalidOperationException("Context already created");
             if (fragment.Data is null || fragment.Data.Length == 0)
                 throw new InvalidOperationException($"Fragment '{fragment.Name}' data is empty");
             if (!FragmentFits(fragment))
@@ -126,98 +71,6 @@ internal unsafe static partial class MockMemorySpace
                 AddHeapFragment(f);
             }
             return this;
-        }
-
-        private HeapFragment CreateContractDescriptor(int jsonLength, int pointerDataCount)
-        {
-            byte[] descriptor = new byte[_targetTestHelpers.ContractDescriptorSize];
-            _targetTestHelpers.ContractDescriptorFill(descriptor, jsonLength, pointerDataCount);
-            return new HeapFragment
-            {
-                Address = ContractDescriptorAddr,
-                Data = descriptor,
-                Name = "ContractDescriptor"
-            };
-        }
-
-        private string MakeContractsJson()
-        {
-            if (_contracts.Count == 0)
-                return string.Empty;
-            StringBuilder sb = new ();
-            foreach (var c in _contracts)
-            {
-                sb.Append($"\"{c}\": 1,");
-            }
-            Debug.Assert(sb.Length > 0);
-            sb.Length--; // remove trailing comma
-            return sb.ToString();
-        }
-
-        private (HeapFragment json, HeapFragment pointerData) CreateDataDescriptor()
-        {
-            string metadataTypesJson = _types is not null ? TargetTestHelpers.MakeTypesJson(_types) : string.Empty;
-            string metadataGlobalsJson = _globals is not null ? TargetTestHelpers.MakeGlobalsJson(_globals) : string.Empty;
-            string interpolatedContracts = _contracts is not null ? MakeContractsJson() : string.Empty;
-            byte[] jsonBytes = Encoding.UTF8.GetBytes($$"""
-            {
-                "version": 0,
-                "baseline": "empty",
-                "contracts": { {{ interpolatedContracts }} },
-                "types": { {{metadataTypesJson}} },
-                "globals": { {{metadataGlobalsJson}} }
-            }
-            """);
-            HeapFragment json = new () {
-                Address = JsonDescriptorAddr,
-                Data = jsonBytes,
-                Name = "JsonDescriptor"
-            };
-
-            HeapFragment pointerData;
-            if (_indirectValues != null)
-            {
-                int pointerSize = _targetTestHelpers.PointerSize;
-                byte[] pointerDataBytes = new byte[_indirectValues.Count * pointerSize];
-                int offset = 0;
-                foreach (var value in _indirectValues)
-                {
-                    _targetTestHelpers.WritePointer(pointerDataBytes.AsSpan(offset, pointerSize), value);
-                    offset += pointerSize;
-                }
-                pointerData =new HeapFragment {
-                    Address = ContractPointerDataAddr,
-                    Data = pointerDataBytes,
-                    Name = "PointerData"
-                };
-            } else {
-                pointerData = new HeapFragment
-                {
-                    Address = ContractPointerDataAddr,
-                    Data = Array.Empty<byte>(),
-                    Name = "PointerData"
-                };
-            }
-            return (json, pointerData);
-        }
-
-        private ulong CreateDescriptorFragments()
-        {
-            if (_created)
-                throw new InvalidOperationException("Context already created");
-            (var json, var pointerData) = CreateDataDescriptor();
-
-            int pointerDataCount = pointerData.Data is null ? 0 : pointerData.Data.Length / _targetTestHelpers.PointerSize;
-
-            HeapFragment descriptor = CreateContractDescriptor(json.Data.Length, pointerDataCount);
-
-            AddHeapFragment(descriptor);
-            AddHeapFragment(json);
-            if (pointerData.Data.Length > 0)
-                AddHeapFragment(pointerData);
-
-            _created = true;
-            return descriptor.Address;
         }
 
         internal ReadContext GetReadContext()
@@ -246,20 +99,9 @@ internal unsafe static partial class MockMemorySpace
             return true;
         }
 
-        public bool TryCreateTarget([NotNullWhen(true)] out ContractDescriptorTarget? target)
-        {
-            if (_created)
-                throw new InvalidOperationException("Context already created");
-            ulong contractDescriptorAddress = CreateDescriptorFragments();
-            ReadContext context = GetReadContext();
-            return ContractDescriptorTarget.TryCreate(contractDescriptorAddress, context.ReadFromTarget, out target);
-        }
-
         // Get an allocator for a range of addresses to simplify creating heap fragments
         public BumpAllocator CreateAllocator(ulong start, ulong end, int minAlign = 16)
         {
-            if (_created)
-                throw new InvalidOperationException("Context already created");
             BumpAllocator allocator = new BumpAllocator(start, end) { MinAlign = minAlign };
             foreach (var a in _allocators)
             {
